Update VC8000 driver module for JPEG decode
Add first-frame check for VC8000 JPEG decode.
The VC8000 JPEG driver determines the end of the stream by detecting a
frame with a size >= 0x200000, which was learned from the MJPEG test
with GStreamer. While this approach works well for motion JPEG cases,
it can result in dropped JPEG decoding when decoding a single picture,
such as in the example code vc8000-jpeg.
This patch excludes the end-of-stream check for the first decode.

Change-Id: Iee9eea9f08ed47291d2c36b25f68e66fc7b7e954